backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
psycopg2-binary==2.9.10
redis[hiredis]==5.2.1
websockets==14.1
pydantic==2.10.4
python-dotenv==1.0.1
httpx==0.28.1
